Refactor binary ops
