+++ prover/src/opcodes/binary/mod.rs
@@ -1,0 +1,7 @@
+mod b128;
+mod b32;
+
+pub use b128::{B128AddTable, B128MulTable};
+pub use b32::{
+    AndTable, AndiTable, B32MulTable, B32MuliTable, OrTable, OriTable, XorTable, XoriTable,
+};
